requirements/developer.txt
-r production.txt

# -----------------------------
# DEVELOPMENT ONLY REQUIREMENTS
# -----------------------------

coloredlogs==7.3.1
django-debug-toolbar==1.9.1
factory-boy==2.9.2
flake8==3.5.0
flake8-blind-except==0.1.1
flake8-coding==1.3.0
flake8-commas==0.4.3
flake8-debugger==3.0.0
flake8-docstrings==1.1.0
flake8-isort==2.2.2
flake8-pep3101==1.1
flake8-quotes==0.12.1
flake8-string-format==0.2.3
freezegun==0.3.9
httmock==1.2.6
hypothesis[django]==3.41.0
ipdb==0.10.3
isort==4.2.15
mccabe==0.6.1
mock==2.0.0
parso==0.1.1
pep257==0.7.0
pep8-naming==0.4.1
pip==8.1.2
pip-check==0.2
pycodestyle==2.3.1
pydocstyle==2.1.1
pyflakes==1.6.0
pytest==3.3.1
pytest-cache==1.0
pytest-cov==2.5.1
pytest-django==3.1.2
radon==2.1.1
